Determine Double-Hit Genes

Determine Double-Hit Genes

determine_double_hit_genes(
  annovar_csv_path,
  scna_df,
  gene_overlap_threshold = 25,
  log2_hom_loss_threshold = -1
)

Arguments

annovar_csv_path
path to 'ANNOVAR' csv output file

scna_df
the SCNA segments data frame. Must contain:

chr

chromosome the segment is located in

start

start position of the segment

end

end position of the segment

log2ratio

log2 ratio of the segment

gene_overlap_threshold
the percentage threshold for the overlap between a segment and a transcript (default = 25). This means that if only a segment overlaps a transcript more than this threshold, the transcript is assigned the segment's SCNA event.

log2_hom_loss_threshold
to determine double-hit events, the log2 threshold for identifying homozygous loss events (default = -1).

Value

vector of gene symbols that are subject to double-hit event(s), i.e. non-synonymous mutation + homozygous CN loss

Examples

path2annovar_csv <- system.file("extdata/example.hg19_multianno.csv",
                                package = "driveR")
if (FALSE) {
dhit_genes <- driveR:::determine_double_hit_genes(path2annovar_csv, example_scna_table)
}